Move the scripts out of the Dockerfile and add deploy
Moving the scripts out of the Dockerfile itself will make it easier to
maintain. The version number is taken care of by the fact that there
is only one antlr jar file (and now also one antlr zip file) in the
container, so a wildcard will only pick up the one we're interested in.

I've also added a deploy script that will deploy the C++ runtime files
to the target directory -- which is useful for projects that generate
C++ code and don't want to lug the runtime around. Also, this means
the runtime and code generator stay in sync.